+++ core/modules/modAeroMigration.class.php
@@ -36,7 +36,7 @@ class modAeroMigration extends DolibarrModules
         $this->module_position = '91';
         $this->name            = preg_replace('/^mod/i', '', get_class($this));
         $this->description     = 'Reprise de données de l\'ancien ERP vers Dolibarr';
-        $this->version         = '0.23.0';
+        $this->version         = '0.24.0';
         $this->const_name      = 'MAIN_MODULE_'.strtoupper($this->name);
         $this->picto           = 'fa-database_fas_#1A3085';